requirement.txt
beautifulsoup4==4.12.2
bs4==0.0.1
Flask==3.0.0
gunicorn==21.2.0
humanize==4.8.0
python-decouple==3.8
selenium==4.15.2
Werkzeug==3.0.1
Flask_Cors==4.0.0